Rip out MobileView.
