+++ app/src/main/java/org/wikipedia/analytics/SessionFunnel.java
@@ -5,7 +5,6 @@ import android.text.format.DateUtils;
 import org.wikipedia.WikipediaApp;
 import org.wikipedia.history.HistoryEntry;
 import org.wikipedia.settings.Prefs;
-import org.wikipedia.settings.RbSwitch;
 import org.wikipedia.util.StringUtil;
 
 public class SessionFunnel extends Funnel {
@@ -113,7 +112,7 @@ public class SessionFunnel extends Funnel {
                 "leadLatency", sessionData.getLeadLatency(),
                 "restLatency", sessionData.getRestLatency(),
                 "languages", StringUtil.listToJsonArrayString(getApp().language().getAppLanguageCodes()),
-                "apiMode", RbSwitch.INSTANCE.isRestBaseEnabled() ? 1 : 0
+                "apiMode", 1
         );
     }
 }
