+++ app/src/main/java/org/wikipedia/analytics/SuggestedPagesFunnel.java
@@ -6,7 +6,7 @@ import androidx.annotation.NonNull;
 
 import org.json.JSONObject;
 import org.wikipedia.WikipediaApp;
-import org.wikipedia.dataclient.restbase.page.RbPageSummary;
+import org.wikipedia.dataclient.page.PageSummary;
 import org.wikipedia.page.PageTitle;
 
 import java.util.List;
@@ -37,7 +37,7 @@ public class SuggestedPagesFunnel extends Funnel {
         this.latency = latency;
     }
 
-    public void logSuggestionsShown(PageTitle currentPageTitle, List<RbPageSummary> suggestedTitles) {
+    public void logSuggestionsShown(PageTitle currentPageTitle, List<PageSummary> suggestedTitles) {
         log(
                 currentPageTitle.getWikiSite(),
                 "action", "shown",
@@ -47,7 +47,7 @@ public class SuggestedPagesFunnel extends Funnel {
         );
     }
 
-    public void logSuggestionClicked(PageTitle currentPageTitle, List<RbPageSummary> suggestedTitles,
+    public void logSuggestionClicked(PageTitle currentPageTitle, List<PageSummary> suggestedTitles,
                                      int clickedIndex) {
         log(
                 currentPageTitle.getWikiSite(),
