+++ app/src/main/java/org/wikipedia/dataclient/page/PageRemaining.java
@@ -1,14 +1,24 @@
 package org.wikipedia.dataclient.page;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import org.wikipedia.page.Section;
 
+import java.util.Collections;
 import java.util.List;
 
 /**
- * Gson POJI for loading remaining page content.
+ * Gson POJO for loading remaining page content.
  */
-public interface PageRemaining {
-    @NonNull List<Section> sections();
+public class PageRemaining {
+    @SuppressWarnings("unused") @Nullable
+    private List<Section> sections;
+
+    @NonNull public List<Section> sections() {
+        if (sections == null) {
+            return Collections.emptyList();
+        }
+        return sections;
+    }
 }
