+++ app/src/main/java/org/wikipedia/dataclient/restbase/RbRelatedPages.java
@@ -3,24 +3,24 @@ package org.wikipedia.dataclient.restbase;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
-import org.wikipedia.dataclient.restbase.page.RbPageSummary;
+import org.wikipedia.dataclient.page.PageSummary;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class RbRelatedPages {
-    @SuppressWarnings("unused") @Nullable private List<RbPageSummary> pages;
+    @SuppressWarnings("unused") @Nullable private List<PageSummary> pages;
 
     @Nullable
-    public List<RbPageSummary> getPages() {
+    public List<PageSummary> getPages() {
         return pages;
     }
 
     @NonNull
-    public List<RbPageSummary> getPages(int limit) {
-        List<RbPageSummary> list = new ArrayList<>();
+    public List<PageSummary> getPages(int limit) {
+        List<PageSummary> list = new ArrayList<>();
         if (getPages() != null) {
-            for (RbPageSummary page : getPages()) {
+            for (PageSummary page : getPages()) {
                 list.add(page);
                 if (limit == list.size()) {
                     break;
