+++ app/src/main/java/org/wikipedia/feed/aggregated/AggregatedFeedContent.java
@@ -4,7 +4,7 @@ import androidx.annotation.Nullable;
 
 import com.google.gson.annotations.SerializedName;
 
-import org.wikipedia.dataclient.restbase.page.RbPageSummary;
+import org.wikipedia.dataclient.page.PageSummary;
 import org.wikipedia.feed.image.FeaturedImage;
 import org.wikipedia.feed.mostread.MostReadArticles;
 import org.wikipedia.feed.news.NewsItem;
@@ -13,7 +13,7 @@ import org.wikipedia.feed.onthisday.OnThisDay;
 import java.util.List;
 
 public class AggregatedFeedContent {
-    @SuppressWarnings("unused") @Nullable private RbPageSummary tfa;
+    @SuppressWarnings("unused") @Nullable private PageSummary tfa;
     @SuppressWarnings("unused") @Nullable private List<NewsItem> news;
     @SuppressWarnings("unused") @SerializedName("mostread") @Nullable private MostReadArticles mostRead;
     @SuppressWarnings("unused") @Nullable private FeaturedImage image;
@@ -25,7 +25,7 @@ public class AggregatedFeedContent {
     }
 
     @Nullable
-    public RbPageSummary tfa() {
+    public PageSummary tfa() {
         return tfa;
     }
