+++ app/src/main/java/org/wikipedia/feed/featured/FeaturedArticleCard.java
@@ -9,7 +9,7 @@ import org.apache.commons.lang3.StringUtils;
 import org.wikipedia.R;
 import org.wikipedia.WikipediaApp;
 import org.wikipedia.dataclient.WikiSite;
-import org.wikipedia.dataclient.restbase.page.RbPageSummary;
+import org.wikipedia.dataclient.page.PageSummary;
 import org.wikipedia.feed.model.CardType;
 import org.wikipedia.feed.model.WikiSiteCard;
 import org.wikipedia.history.HistoryEntry;
@@ -17,10 +17,10 @@ import org.wikipedia.page.PageTitle;
 import org.wikipedia.util.DateUtil;
 
 public class FeaturedArticleCard extends WikiSiteCard {
-    @NonNull private RbPageSummary page;
+    @NonNull private PageSummary page;
     private int age;
 
-    public FeaturedArticleCard(@NonNull RbPageSummary page, int age, @NonNull WikiSite wiki) {
+    public FeaturedArticleCard(@NonNull PageSummary page, int age, @NonNull WikiSite wiki) {
         super(wiki);
         this.page = page;
         this.age = age;
