+++ app/src/main/java/org/wikipedia/feed/mostread/MostReadArticles.java
@@ -2,7 +2,7 @@ package org.wikipedia.feed.mostread;
 
 import androidx.annotation.NonNull;
 
-import org.wikipedia.dataclient.restbase.page.RbPageSummary;
+import org.wikipedia.dataclient.page.PageSummary;
 import org.wikipedia.json.annotations.Required;
 
 import java.util.Date;
@@ -10,13 +10,13 @@ import java.util.List;
 
 public final class MostReadArticles {
     @SuppressWarnings("unused,NullableProblems") @Required @NonNull private Date date;
-    @SuppressWarnings("unused,NullableProblems") @Required @NonNull private List<RbPageSummary> articles;
+    @SuppressWarnings("unused,NullableProblems") @Required @NonNull private List<PageSummary> articles;
 
     @NonNull public Date date() {
         return date;
     }
 
-    @NonNull public List<RbPageSummary> articles() {
+    @NonNull public List<PageSummary> articles() {
         return articles;
     }
 }
