+++ app/src/main/java/org/wikipedia/feed/news/NewsItem.java
@@ -7,7 +7,7 @@ import androidx.annotation.Nullable;
 
 import org.apache.commons.lang3.StringUtils;
 import org.wikipedia.dataclient.WikiSite;
-import org.wikipedia.dataclient.restbase.page.RbPageSummary;
+import org.wikipedia.dataclient.page.PageSummary;
 import org.wikipedia.json.annotations.Required;
 
 import java.util.ArrayList;
@@ -19,20 +19,20 @@ import static org.wikipedia.util.ImageUrlUtil.getUrlForSize;
 
 public final class NewsItem {
     @SuppressWarnings("unused") @Required @Nullable private String story;
-    @SuppressWarnings("unused") @Nullable private List<RbPageSummary> links
+    @SuppressWarnings("unused") @Nullable private List<PageSummary> links
             = Collections.emptyList();
 
     @NonNull String story() {
         return StringUtils.defaultString(story);
     }
 
-    @NonNull public List<RbPageSummary> links() {
+    @NonNull public List<PageSummary> links() {
         return links != null ? links : Collections.emptyList();
     }
 
     @NonNull List<NewsLinkCard> linkCards(WikiSite wiki) {
         List<NewsLinkCard> linkCards = new ArrayList<>();
-        for (RbPageSummary link : links()) {
+        for (PageSummary link : links()) {
             if (link == null) {
                 continue;
             }
@@ -54,8 +54,8 @@ public final class NewsItem {
      * Iterate through the CardPageItems associated with the news story's links and return the first
      * thumb URI found.
      */
-    @Nullable private Uri getFirstImageUri(List<RbPageSummary> links) {
-        for (RbPageSummary link : links) {
+    @Nullable private Uri getFirstImageUri(List<PageSummary> links) {
+        for (PageSummary link : links) {
             if (link == null) {
                 continue;
             }
