+++ app/src/main/java/org/wikipedia/feed/news/NewsItemCard.java
@@ -9,7 +9,7 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import org.wikipedia.dataclient.WikiSite;
-import org.wikipedia.dataclient.restbase.page.RbPageSummary;
+import org.wikipedia.dataclient.page.PageSummary;
 import org.wikipedia.feed.model.Card;
 import org.wikipedia.feed.model.CardType;
 import org.wikipedia.richtext.RichTextUtil;
@@ -47,7 +47,7 @@ public class NewsItemCard extends Card {
         return removeImageCaption(StringUtil.fromHtml(newsItem.story()));
     }
 
-    @NonNull public List<RbPageSummary> links() {
+    @NonNull public List<PageSummary> links() {
         return newsItem.links();
     }
