+++ app/src/main/java/org/wikipedia/page/Page.java
@@ -2,9 +2,6 @@ package org.wikipedia.page;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.annotation.VisibleForTesting;
-
-import org.wikipedia.settings.RbSwitch;
 
 import java.util.List;
 
@@ -12,46 +9,15 @@ import java.util.List;
  * Represents a particular page along with its full contents.
  */
 public class Page {
-    @VisibleForTesting static final int MEDIAWIKI_ORIGIN = 0;
-    @VisibleForTesting static final int RESTBASE_ORIGIN = 1;
-
     @NonNull private final PageTitle title;
     @NonNull private final List<Section> sections;
     @NonNull private final PageProperties pageProperties;
 
-    /**
-     * An indicator what payload version the page content was originally retrieved from.
-     * If it's set to RESTBASE_ORIGIN the it came from the Mobile Content Service
-     * (via RESTBase). This is esp. useful for saved pages, so that an older saved page will get the
-     * correct kind of DOM transformations applied.
-     */
-    private int version = MEDIAWIKI_ORIGIN;
-
     /** Regular constructor */
-    public Page(@NonNull PageTitle title, @NonNull List<Section> sections,
-                @NonNull PageProperties pageProperties) {
-        if (RbSwitch.INSTANCE.isRestBaseEnabled(title.getWikiSite())) {
-            this.version = RESTBASE_ORIGIN;
-        }
+    public Page(@NonNull PageTitle title, @NonNull List<Section> sections, @NonNull PageProperties pageProperties) {
         this.title = title;
         this.sections = sections;
         this.pageProperties = pageProperties;
-    }
-
-    @VisibleForTesting Page(@NonNull PageTitle title, @NonNull List<Section> sections,
-         @NonNull PageProperties pageProperties, int version) {
-        this.version = version;
-        this.title = title;
-        this.sections = sections;
-        this.pageProperties = pageProperties;
-    }
-
-    /**
-     * This could also be called getVersion but since there are only two different versions
-     * I like to call it isFromRestBase to make it clearer.
-     */
-    public boolean isFromRestBase() {
-        return version == RESTBASE_ORIGIN;
     }
 
     @NonNull public PageTitle getTitle() {
