+++ app/src/main/java/org/wikipedia/page/PageCacher.java
@@ -5,8 +5,7 @@ import android.annotation.SuppressLint;
 import androidx.annotation.NonNull;
 
 import org.wikipedia.WikipediaApp;
-import org.wikipedia.dataclient.page.PageClient;
-import org.wikipedia.dataclient.page.PageClientFactory;
+import org.wikipedia.dataclient.ServiceFactory;
 import org.wikipedia.dataclient.page.PageLead;
 import org.wikipedia.dataclient.page.PageRemaining;
 import org.wikipedia.util.log.L;
@@ -16,15 +15,12 @@ import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.schedulers.Schedulers;
 import retrofit2.Response;
 
-import static org.wikipedia.util.DimenUtil.calculateLeadImageWidth;
-
 final class PageCacher {
 
     @SuppressLint("CheckResult")
     static void loadIntoCache(@NonNull PageTitle title) {
         L.d("Loading page into cache: " + title.getPrefixedText());
-        PageClient client = PageClientFactory.create(title.getWikiSite(), title.namespace());
-        Observable.zip(leadReq(client, title), remainingReq(client, title), (leadRsp, sectionsRsp) -> leadRsp)
+        Observable.zip(leadReq(title), remainingReq(title), (leadRsp, sectionsRsp) -> leadRsp)
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
                 .subscribe(leadRsp -> {
@@ -39,15 +35,13 @@ final class PageCacher {
     }
 
     @NonNull
-    private static Observable<Response<PageLead>> leadReq(@NonNull PageClient client, @NonNull PageTitle title) {
-        return client.lead(title.getWikiSite(), null, null, null, title.getPrefixedText(),
-                calculateLeadImageWidth());
+    private static Observable<Response<PageLead>> leadReq(@NonNull PageTitle title) {
+        return ServiceFactory.getRest(title.getWikiSite()).getLeadSection(null, null, null, title.getPrefixedText());
     }
 
     @NonNull
-    private static Observable<Response<PageRemaining>> remainingReq(@NonNull PageClient client,
-                                                    @NonNull PageTitle title) {
-        return client.sections(title.getWikiSite(), null, null, title.getPrefixedText());
+    private static Observable<Response<PageRemaining>> remainingReq(@NonNull PageTitle title) {
+        return ServiceFactory.getRest(title.getWikiSite()).getRemainingSections(null, null, title.getPrefixedText());
     }
 
     private PageCacher() { }
