+++ app/src/main/java/org/wikipedia/suggestededits/provider/SuggestedEditItem.kt
@@ -1,11 +1,11 @@
 package org.wikipedia.suggestededits.provider
 
 import com.google.gson.annotations.SerializedName
-import org.wikipedia.dataclient.restbase.page.RbPageSummary
+import org.wikipedia.dataclient.page.PageSummary
 import org.wikipedia.gallery.GalleryItem
 import org.wikipedia.wikidata.Entities
 
-class SuggestedEditItem : RbPageSummary() {
+class SuggestedEditItem : PageSummary() {
     @SerializedName("structured") private val structuredData: GalleryItem.StructuredData? = null
     @SerializedName("wikibase_item") val entity: Entities.Entity? = null
