+++ app/src/main/java/org/wikipedia/widgets/WidgetProviderFeaturedPage.java
@@ -18,15 +18,13 @@ import org.wikipedia.Constants;
 import org.wikipedia.R;
 import org.wikipedia.WikipediaApp;
 import org.wikipedia.dataclient.ServiceFactory;
-import org.wikipedia.dataclient.page.PageClientFactory;
 import org.wikipedia.dataclient.page.PageLead;
-import org.wikipedia.dataclient.restbase.page.RbPageSummary;
+import org.wikipedia.dataclient.page.PageSummary;
 import org.wikipedia.feed.model.UtcDate;
 import org.wikipedia.page.PageActivity;
 import org.wikipedia.page.PageTitle;
 import org.wikipedia.staticdata.MainPageNameData;
 import org.wikipedia.util.DateUtil;
-import org.wikipedia.util.DimenUtil;
 import org.wikipedia.util.StringUtil;
 import org.wikipedia.util.log.L;
 
@@ -102,8 +100,8 @@ public class WidgetProviderFeaturedPage extends AppWidgetProvider {
                         return Observable.just(response.tfa());
                     } else {
                         // TODO: this logic can be removed if the feed API can return the featured article for all languages.
-                        return PageClientFactory.create(mainPageTitle.getWikiSite(), mainPageTitle.namespace())
-                                .lead(mainPageTitle.getWikiSite(), null, null, null, mainPageTitle.getPrefixedText(), DimenUtil.calculateLeadImageWidth());
+                        return ServiceFactory.getRest(mainPageTitle.getWikiSite()).getLeadSection(null,
+                                null, null, mainPageTitle.getPrefixedText());
                     }
                 })
                 .subscribeOn(Schedulers.io())
@@ -116,8 +114,8 @@ public class WidgetProviderFeaturedPage extends AppWidgetProvider {
                     return Observable.just(response);
                 })
                 .subscribe(response -> {
-                    CharSequence widgetText = StringUtil.fromHtml(((RbPageSummary) response).getDisplayTitle());
-                    PageTitle pageTitle = ((RbPageSummary) response).getPageTitle(app.getWikiSite());
+                    CharSequence widgetText = StringUtil.fromHtml(((PageSummary) response).getDisplayTitle());
+                    PageTitle pageTitle = ((PageSummary) response).getPageTitle(app.getWikiSite());
                     cb.onFeaturedArticleReceived(pageTitle, widgetText);
                 }, throwable -> {
                     cb.onFeaturedArticleReceived(mainPageTitle, mainPageTitle.getDisplayText());
